lambda-load-old-data/requirements.txt
pylint
pytest
pandas
sqlalchemy
boto3
python-dotenv